Added some error handling, should the bufkit decoder fail to find data.  Also fixed the "program explodes when data are missing" issue.  This involved moving all the data download code from SPCWindow.py to full_gui.py, which simplifies what happens if it can't find the data.
